added search and dragable list

+++ app/src/main/java/com/moor/scannit/ui/home/HomeFragment.kt
@@ -3,8 +3,8 @@ package com.moor.scannit.ui.home
 
 import android.Manifest
 import android.app.Activity
-import android.app.Activity.RESULT_OK
-import android.app.AlertDialog
+import android.app.SearchManager
+import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.graphics.BitmapFactory
@@ -12,7 +12,7 @@ import android.net.Uri
 import android.os.Bundle
 import android.provider.MediaStore
 import android.view.*
-import android.widget.EditText
+import androidx.appcompat.widget.SearchView
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
@@ -20,7 +20,9 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.moor.scannit.R
 import com.moor.scannit.RESULT_LOAD_IMAGE
@@ -29,6 +31,7 @@ import com.moor.scannit.databinding.FragmentHomeBinding
 import com.moor.scannit.getRealPathFromUri
 import com.moor.scannit.ui.AdapterCallback
 import com.moor.scannit.ui.camera.CameraViewModel
+import java.util.*
 
 
 class HomeFragment : Fragment(), AdapterCallback<Document> {
@@ -53,14 +56,17 @@ class HomeFragment : Fragment(), AdapterCallback<Document> {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
+
+
         binding = FragmentHomeBinding.inflate(inflater,container,false).apply {
             folderListView.apply {
                 adapter= documentAdapter.apply {
                     listener= this@HomeFragment
                 }
-                //layoutManager= GridLayoutManager(context,3)
+
                 layoutManager = LinearLayoutManager(requireContext())
                 addItemDecoration(DividerItemDecoration(context,DividerItemDecoration.VERTICAL))
+                helper.attachToRecyclerView(this)
             }
             scanButton.setOnClickListener {
                 val action = HomeFragmentDirections.actionHomeFragmentToCameraFragment(0)
@@ -102,6 +108,21 @@ class HomeFragment : Fragment(), AdapterCallback<Document> {
 
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
         inflater.inflate(R.menu.menu_home,menu)
+        val searchManager = context?.getSystemService(Context.SEARCH_SERVICE) as SearchManager
+        (menu.findItem(R.id.search).actionView as SearchView).apply {
+            setSearchableInfo(searchManager.getSearchableInfo(activity?.componentName))
+            setOnQueryTextListener(object :SearchView.OnQueryTextListener{
+                override fun onQueryTextSubmit(query: String?): Boolean {
+                    return true
+                }
+
+                override fun onQueryTextChange(text: String): Boolean {
+                    viewModel.filterDocuments(text)
+                    return true
+                }
+
+            })
+        }
         super.onCreateOptionsMenu(menu, inflater)
 
     }
@@ -124,17 +145,21 @@ class HomeFragment : Fragment(), AdapterCallback<Document> {
                        when(item.itemId){
                            R.id.action_delete->{
                                val ids= documentAdapter.selectedIds
-                               val builder = MaterialAlertDialogBuilder(requireContext())
-                               builder.setMessage("Are you sure you want to delete ${ids.size} document(s) ?")
-                               builder.setCancelable(false)
-                               builder.setPositiveButton("Yes"){ dialog, which ->
-                                   viewModel.deleteDocuments(ids.toLongArray())
-                               }
-                               builder.setNegativeButton("No",{_,a->})
-                               builder.setOnDismissListener {
+                               if (ids.any()){
+                                   val builder = MaterialAlertDialogBuilder(requireContext())
+                                   builder.setMessage("Are you sure you want to delete ${ids.size} document(s) ?")
+                                   builder.setCancelable(false)
+                                   builder.setPositiveButton("Yes"){ dialog, which ->
+                                       viewModel.deleteDocuments(ids.toLongArray())
+                                   }
+                                   builder.setNegativeButton("No",{_,a->})
+                                   builder.setOnDismissListener {
+                                       callback.finishActionMode()
+                                   }
+                                   builder.show()
+                               }else{
                                    callback.finishActionMode()
                                }
-                               builder.show()
 
                            }
                        }
@@ -160,8 +185,6 @@ class HomeFragment : Fragment(), AdapterCallback<Document> {
         findNavController().navigate(action)
     }
 
-
-
     private  fun pickImage(){
         val i = Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI)
         startActivityForResult(i, RESULT_LOAD_IMAGE)
@@ -181,5 +204,31 @@ class HomeFragment : Fragment(), AdapterCallback<Document> {
     private fun allPermissionsGranted() = REQUIRED_PERMISSIONS.all {
         ContextCompat.checkSelfPermission(requireContext(), it) == PackageManager.PERMISSION_GRANTED
     }
+
+    private val helper=ItemTouchHelper(object :ItemTouchHelper.Callback(){
+        override fun getMovementFlags(
+            recyclerView: RecyclerView,
+            viewHolder: RecyclerView.ViewHolder
+        ): Int {
+            return makeMovementFlags(ItemTouchHelper.UP or ItemTouchHelper.DOWN,0)
+        }
+
+        override fun onMove(
+            recyclerView: RecyclerView,
+            dragged: RecyclerView.ViewHolder,
+            target: RecyclerView.ViewHolder
+        ): Boolean {
+            val targetPosition= target.adapterPosition
+            val draggedPosition= dragged.adapterPosition
+
+            Collections.swap(documents,draggedPosition,targetPosition)
+            documentAdapter.notifyItemMoved(draggedPosition,targetPosition)
+            return true
+        }
+
+        override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
+
+        }
+    })
 
 }
